refactor: move test files into __tests__/ subdirectories
